making password-reset send email config
