+++ server/middlewares/authMiddleware.js
@@ -1,29 +1,13 @@
 const jwt = require('jsonwebtoken');
 const redis_client = require('../config/redisConfig');
 const { isEmpty } = require('../utilities/common');
-
+const { validatePassword, validateEmail } = require('../utilities/validation');
 function verifyToken(req, res, next) {
     try {
         // Bearer token string
         const token = req.headers.authorization.split(' ')[1];
         console.log(token);
-        const decoded = jwt.verify(token, process.env.JWT_ACCESS_SECRET);
-        req.userData = decoded;
-
-        req.token = token;
-
-        // verify blacklisted access token.
-        redis_client.get('BL_' + decoded.sub.toString(), (err, data) => {
-            if (err) throw err;
-
-            if (data === token) {
-                next(
-                    res
-                        .status(401)
-                        .json({ status: false, message: 'blacklisted token.' }),
-                );
-            }
-        });
+        req.userData = jwt.verify(token, process.env.JWT_ACCESS_SECRET);
     } catch (error) {
         next(
             res.status(401).json({
@@ -84,14 +68,7 @@ function verifyRefreshToken(req, res, next) {
     next();
 }
 
-function validateEmail(email) {
-    const re =
-        /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-    return re.test(email.toLowerCase());
-}
-
 function validateInputData(req, res, next) {
-    const re = /\d/;
     const { password, username, email } = req.body;
     let errors = {};
 
@@ -102,7 +79,7 @@ function validateInputData(req, res, next) {
     if (!validateEmail(email)) {
         errors.errorEmail = 'The provided email is not valid.';
     }
-    if (password.length < 6 || password.length > 30 || !re.test(password)) {
+    if (validatePassword(password)) {
         errors.errorPassword =
             'The password must have one digit at least, and to be between 6 and 30 symbols';
     }
