+++ server/routes/authRoute.js
@@ -9,17 +9,17 @@ router.post('/', auth_middleware.verifyToken, (req, res) => {
 router.post(
     '/register',
     auth_middleware.validateInputData,
-    user_controller.Register,
+    user_controller.register,
 );
 
-router.post('/login', user_controller.Login);
+router.post('/login', user_controller.login);
 
 router.post(
     '/token',
     auth_middleware.verifyRefreshToken,
-    user_controller.GetAccess,
+    user_controller.getAccess,
 );
 
-router.post('/logout', auth_middleware.verifyToken, user_controller.Logout);
+router.post('/logout', auth_middleware.verifyToken, user_controller.logout);
 
 module.exports = router;
