+++ server/utilities/validation.js
@@ -1,0 +1,15 @@
+function validateEmail(email) {
+    const re =
+        /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+    return re.test(email.toLowerCase());
+}
+
+function validatePassword(password) {
+    const re = /\d/;
+    return password.length > 10 && password.length < 35 && re.test(password);
+}
+
+module.exports = {
+    validateEmail,
+    validatePassword,
+};
